Replace header text with header png of text.
I know that images of text are relatively evil, however browsershots.org indicates to me that the vast majority of browsers on the vast majority of platforms are completely incapable of rendering italic text in any way other than terribly (eg, no anti-aliasing, so it looks very blocky and jaggy). On the other hand, very few browsers have difficulty with transparent PNGs, so this seems like a good tradeoff. I even added alt text for browsers that can't display images.